Decode into interface fields that are initialized pointers
Decoder can now handle interface fields in a correct way.
If an interface field was initialized with a settable pointer value
it will try to decode into that type instead of setting the interface
to string. Nil interface, or interface with a non pointer value, is still
being decoded to string.